requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
auth==0.5.3
blinker==1.9.0
boto3==1.37.30
botocore==1.37.30
cachetools==5.5.2
category_encoders==2.7.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
dash==3.0.2
databricks-sdk==0.49.0
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.1
eventlet==0.39.1
exceptiongroup==1.2.2
executing==2.2.0
falcon==4.0.2
fastapi==0.115.12
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
gotrue==2.12.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.1.0
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mlflow==2.21.3
mlflow-skinny==2.21.3
mongoengine==0.29.1
multidict==6.3.2
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.0
numpy==1.26.4
openai==1.70.0
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
orjson==3.10.16
packaging==24.2
pandas==2.1.4
parso==0.8.4
passlib==1.7.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==5.24.1
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
postgrest==1.0.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycaret==3.3.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
pyod==2.0.4
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
storage3==0.11.3
StrEnum==0.4.15
stripe==12.0.0
supabase==2.15.0
supafunc==0.9.4
tbats==1.1.3
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
wurlitzer==3.1.1
xxhash==3.5.0
yarl==1.19.0
yellowbrick==1.5
zipp==3.21.0

sagemaker~=2.244.2
FLAML~=2.3.5